+++ CHANGELOG.md
@@ -3,15 +3,16 @@
 
 ## 1.1.0
 
-> Released 25 Oct 2024
+> Released 04 Nov 24
 
 Added functions:
 
 * `Get-CEventLog`, as a drop in replacement for using `Get-EventLog` in PowerShell 7.
-* `Install-CEventLog`, as an idempotent version of `Get-CEventLog`.
+* `Install-CEventLog`,  which creates an event log only if it doesn't already exist.
 * `New-CEventLog`, as a drop in replacement for using `New-EventLog` in PowerShell 7.
 * `Remove-CEventLog`, as a drop in replacement for using `Remove-EventLog` in PowerShell 7.
-* `Uninstall-CEventLog`, as an idempotent version of `Remove-EventLog`
+* `Test-CEventLog`, which checks if an event log exists.
+* `Uninstall-CEventLog`, which creates an event log only if it doesn't already exist.
 
 ## 1.0.0